+++ frontend/src/app/main/data/dashboard/shortcuts.cljs
@@ -13,13 +13,16 @@
    [app.main.data.shortcuts :as ds]
    [app.main.store :as st]))
 
+;; Shortcuts definitions
 (def shortcuts
-  {:go-to-search       {:tooltip (ds/meta "F")
-                        :command (ds/c-mod "f")
-                        :subsections [:navigation-dashboard]
-                        :fn #(st/emit! (dcm/go-to-dashboard-search))}
+  {:toggle-theme    {:tooltip (ds/alt "M")
+                     :command (ds/a-mod "m")
+                     :subsections [:general-dashboard]
+                     :fn #(st/emit! (with-meta (du/toggle-theme)
+                                      {::ev/origin "dashboard:shortcuts"}))}})
 
-   :go-to-drafts       {:tooltip "G D"
+(def shortcuts-sidebar-navigation
+  {:go-to-drafts       {:tooltip "G D"
                         :command "g d"
                         :subsections [:navigation-dashboard]
                         :fn #(st/emit! (dcm/go-to-dashboard-files :project-id :default))}
@@ -27,20 +30,35 @@
    :go-to-libs         {:tooltip "G L"
                         :command "g l"
                         :subsections [:navigation-dashboard]
-                        :fn #(st/emit! (dcm/go-to-dashboard-libraries))}
+                        :fn #(st/emit! (dcm/go-to-dashboard-libraries))}})
 
-   :create-new-project {:tooltip "+"
+(def shortcut-search
+  {:go-to-search       {:tooltip (ds/meta "F")
+                        :command (ds/c-mod "f")
+                        :subsections [:navigation-dashboard]
+                        :fn #(st/emit! (dcm/go-to-dashboard-search))}})
+
+(def shortcut-create-new-project
+  {:create-new-project {:tooltip "+"
                         :command "+"
                         :subsections [:general-dashboard]
-                        :fn #(st/emit! (dd/create-element))}
+                        :fn #(st/emit! (dd/create-element))}})
 
-   :toggle-theme    {:tooltip (ds/alt "M")
-                     :command (ds/a-mod "m")
-                     :subsections [:general-dashboard]
-                     :fn #(st/emit! (with-meta (du/toggle-theme)
-                                      {::ev/origin "dashboard:shortcuts"}))}})
+;; Shortcuts combinations for files, drafts, libraries and fonts sections
+(def shortcuts-dashboard
+  (merge shortcuts
+         shortcuts-sidebar-navigation))
 
+(def shortcuts-projects
+  (merge shortcuts
+         shortcuts-sidebar-navigation
+         shortcut-search
+         shortcut-create-new-project))
 
+(def shortcuts-drafts-libraries
+  (merge shortcuts
+         shortcuts-sidebar-navigation
+         shortcut-search))
 
 (defn get-tooltip [shortcut]
   (assert (contains? shortcuts shortcut) (str shortcut))